requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
idna==3.6
pillow==10.2.0
pypng==0.20220715.0
PyQRCode==1.2.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
